requirements.txt
pandas>=2.0.0
streamlit>=1.28.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
